refactor(storage): extract projection read-model views
