+++ framework/core/architecture/ARCHITECTURE_INDEX.md
@@ -24,7 +24,7 @@ Generated from `layers.json` and `build-capabilities.json`. Do not edit by hand.
 | `runtime-ledger-services` | `core/runtime-ledger` | `core/architecture` | `src/libkungfu/src/runtime/durable_ingest.cpp`<br>`src/libkungfu/src/runtime/facts/fact_admission.cpp` | `kungfu_ledger_services`<br>`kungfu` | `embedded-sqlite`<br>`full`<br>`server` | `kungfu_durable_ingest_tests`<br>`kungfu_crash_recovery_tests` | — |
 | `runtime-state-query-services` | `core/runtime-state-query` | `core/architecture` | `src/libkungfu/src/runtime/state_service.cpp`<br>`src/libkungfu/src/runtime/query/fact_query.cpp` | `kungfu_state_query_services`<br>`kungfu_state_cache_services`<br>`kungfu` | `embedded-sqlite`<br>`full`<br>`server` | `kungfu_state_service_contract_tests`<br>`kungfu_bounded_sql_parser_tests` | `storage-query-or-state-failure` |
 | `runtime-live-services` | `core/runtime-live` | `core/architecture` | `src/libkungfu/src/runtime/live/reactor.cpp`<br>`src/libkungfu/src/runtime/live/coordinator.cpp` | `kungfu_live_services`<br>`kungfu` | `embedded-sqlite`<br>`full`<br>`server` | `kungfu_peer_continuity_tests` | `live-continuity-failure` |
-| `runtime-storage-services` | `core/runtime-storage` | `core/architecture` | `src/libkungfu/src/runtime/storage/service.cpp`<br>`src/libkungfu/src/runtime/storage/service_facade.cpp`<br>`src/libkungfu/src/runtime/storage/service_lifecycle.cpp`<br>`src/libkungfu/src/runtime/storage/service_projection.cpp`<br>`src/libkungfu/src/runtime/storage/maintenance_service.cpp` | `kungfu_storage_services`<br>`kungfu` | `embedded-sqlite`<br>`full`<br>`server` | `kungfu_durability_contract_tests`<br>`kungfu_fact_authority_contract_tests`<br>`kungfu_offhost_backup_fixture` | — |
+| `runtime-storage-services` | `core/runtime-storage` | `core/architecture` | `src/libkungfu/src/runtime/storage/service.cpp`<br>`src/libkungfu/src/runtime/storage/service_facade.cpp`<br>`src/libkungfu/src/runtime/storage/service_lifecycle.cpp`<br>`src/libkungfu/src/runtime/storage/service_projection.cpp`<br>`src/libkungfu/src/runtime/storage/service_views.cpp`<br>`src/libkungfu/src/runtime/storage/maintenance_service.cpp` | `kungfu_storage_services`<br>`kungfu` | `embedded-sqlite`<br>`full`<br>`server` | `kungfu_durability_contract_tests`<br>`kungfu_fact_authority_contract_tests`<br>`kungfu_offhost_backup_fixture` | — |
 | `runtime-extension-services` | `core/runtime-extension` | `core/architecture` | `src/libkungfu/src/runtime/kfx/native_registry.cpp`<br>`src/libkungfu/src/runtime/trust/assessment_runtime.cpp` | `kungfu_extension_services`<br>`kungfu` | `embedded-sqlite`<br>`full`<br>`server` | `kungfu_native_kfx_contract_tests`<br>`kungfu_profile_lifecycle_tests`<br>`kungfu_action_geometry_tests`<br>`kungfu_domain_profile_tests`<br>`kungfu_profile_action_tests`<br>`kungfu_action_runtime_tests` | — |
 | `runtime-storage-adapters` | `core/runtime-storage-adapters` | `core/architecture` | `src/libkungfu/src/runtime/storage/provider.cpp` | `kungfu_storage_adapters`<br>`kungfu` | `embedded-sqlite`<br>`full`<br>`server` | `kungfu_durability_contract_tests` | `storage-provider-failure` |
 | `runtime-platform-adapters` | `core/runtime-platform-adapters` | `core/architecture` | `src/libkungfu/src/view/schema.cpp`<br>`src/libkungfu/src/runtime/io/io.cpp` | `kungfu_view_adapters`<br>`kungfu_platform_adapters`<br>`kungfu` | `embedded-sqlite`<br>`full`<br>`server` | `kungfu_view_component_link_tests` | — |
