+++ framework/core/architecture/LAYERS.md
@@ -46,7 +46,7 @@ The production graph is budgeted to 6-12 bounded components.
 | `runtime-ledger-services` | `application-services` | `core/runtime-ledger` | 10 | `kungfu_ledger_services`<br>`kungfu` | `kungfu_durable_ingest_tests`<br>`kungfu_crash_recovery_tests` | `src/libkungfu/src/runtime/durable_ingest.cpp`<br>`src/libkungfu/src/runtime/facts/fact_admission.cpp` |
 | `runtime-state-query-services` | `application-services` | `core/runtime-state-query` | 10 | `kungfu_state_query_services`<br>`kungfu_state_cache_services`<br>`kungfu` | `kungfu_state_service_contract_tests`<br>`kungfu_bounded_sql_parser_tests` | `src/libkungfu/src/runtime/state_service.cpp`<br>`src/libkungfu/src/runtime/query/fact_query.cpp` |
 | `runtime-live-services` | `application-services` | `core/runtime-live` | 6 | `kungfu_live_services`<br>`kungfu` | `kungfu_peer_continuity_tests` | `src/libkungfu/src/runtime/live/reactor.cpp`<br>`src/libkungfu/src/runtime/live/coordinator.cpp` |
-| `runtime-storage-services` | `application-services` | `core/runtime-storage` | 32 | `kungfu_storage_services`<br>`kungfu` | `kungfu_durability_contract_tests`<br>`kungfu_fact_authority_contract_tests`<br>`kungfu_offhost_backup_fixture` | `src/libkungfu/src/runtime/storage/service.cpp`<br>`src/libkungfu/src/runtime/storage/service_facade.cpp`<br>`src/libkungfu/src/runtime/storage/service_lifecycle.cpp`<br>`src/libkungfu/src/runtime/storage/service_projection.cpp`<br>`src/libkungfu/src/runtime/storage/maintenance_service.cpp` |
+| `runtime-storage-services` | `application-services` | `core/runtime-storage` | 33 | `kungfu_storage_services`<br>`kungfu` | `kungfu_durability_contract_tests`<br>`kungfu_fact_authority_contract_tests`<br>`kungfu_offhost_backup_fixture` | `src/libkungfu/src/runtime/storage/service.cpp`<br>`src/libkungfu/src/runtime/storage/service_facade.cpp`<br>`src/libkungfu/src/runtime/storage/service_lifecycle.cpp`<br>`src/libkungfu/src/runtime/storage/service_projection.cpp`<br>`src/libkungfu/src/runtime/storage/service_views.cpp`<br>`src/libkungfu/src/runtime/storage/maintenance_service.cpp` |
 | `runtime-extension-services` | `application-services` | `core/runtime-extension` | 16 | `kungfu_extension_services`<br>`kungfu` | `kungfu_native_kfx_contract_tests`<br>`kungfu_profile_lifecycle_tests`<br>`kungfu_action_geometry_tests`<br>`kungfu_domain_profile_tests`<br>`kungfu_profile_action_tests`<br>`kungfu_action_runtime_tests` | `src/libkungfu/src/runtime/kfx/native_registry.cpp`<br>`src/libkungfu/src/runtime/trust/assessment_runtime.cpp` |
 | `runtime-storage-adapters` | `adapters` | `core/runtime-storage-adapters` | 7 | `kungfu_storage_adapters`<br>`kungfu` | `kungfu_durability_contract_tests` | `src/libkungfu/src/runtime/storage/provider.cpp` |
 | `runtime-platform-adapters` | `adapters` | `core/runtime-platform-adapters` | 10 | `kungfu_view_adapters`<br>`kungfu_platform_adapters`<br>`kungfu` | `kungfu_view_component_link_tests` | `src/libkungfu/src/view/schema.cpp`<br>`src/libkungfu/src/runtime/io/io.cpp` |
@@ -66,7 +66,7 @@ from the same authority as this map.
 | `kungfu_state_query_services` | `OBJECT` | `runtime-state-query-services` | `kungfu_contracts` | 6 |
 | `kungfu_state_cache_services` | `OBJECT` | `runtime-state-query-services` | `kungfu_contracts` | 3 |
 | `kungfu_live_services` | `OBJECT` | `runtime-live-services` | `kungfu_contracts` | 6 |
-| `kungfu_storage_services` | `OBJECT` | `runtime-storage-services` | `kungfu_contracts` | 27 |
+| `kungfu_storage_services` | `OBJECT` | `runtime-storage-services` | `kungfu_contracts` | 28 |
 | `kungfu_extension_services` | `OBJECT` | `runtime-extension-services` | `kungfu_contracts` | 14 |
 | `kungfu_storage_adapters` | `OBJECT` | `runtime-storage-adapters` | `kungfu_storage_services`<br>`kungfu_contracts` | 6 |
 | `kungfu_view_adapters` | `OBJECT` | `runtime-platform-adapters` | `kungfu_contracts` | 2 |
@@ -81,7 +81,8 @@ back into the compatibility facade.
 
 | Responsibility | Source | Line budget |
 | --- | --- | ---: |
-| Storage projection, qualification and journal-derived shared helpers | `src/libkungfu/src/runtime/storage/service.cpp` | 1600 |
+| Storage qualification and journal-derived shared helpers | `src/libkungfu/src/runtime/storage/service.cpp` | 1050 |
+| Workspace layout and storage projection read-model rendering | `src/libkungfu/src/runtime/storage/service_views.cpp` | 260 |
 | Stable storage JSON edge and compatibility facade | `src/libkungfu/src/runtime/storage/service_facade.cpp` | 1000 |
 | Typed storage service composition and lifecycle delegation | `src/libkungfu/src/runtime/storage/service_lifecycle.cpp` | 320 |
 | Episode projection and stable JSON rendering | `src/libkungfu/src/runtime/storage/service_projection.cpp` | 360 |
@@ -143,6 +144,7 @@ authority; this architecture contract does not own a second ledger.
 | Live peer coordination and continuity | `runtime-live-services` → `src/libkungfu/src/runtime/live/coordinator.cpp` |
 | KFX, profile lifecycle or trust assessment | `runtime-extension-services` → `src/libkungfu/src/runtime/kfx/native_registry.cpp` |
 | Storage facade or shared application helpers | `runtime-storage-services` → `src/libkungfu/src/runtime/storage/service_facade.cpp` |
+| Workspace layout or storage projection read-model rendering | `runtime-storage-services` → `src/libkungfu/src/runtime/storage/service_views.cpp` |
 | Storage status, fsck, rebuild, GC or compact planning | `runtime-storage-services` → `src/libkungfu/src/runtime/storage/maintenance_service.cpp` |
 | Storage bundle import, export or sync verification | `runtime-storage-services` → `src/libkungfu/src/runtime/storage/transfer_service.cpp` |
 | Storage provider registry or RocksDB adapter integration | `runtime-storage-adapters` → `src/libkungfu/src/runtime/storage/provider.cpp` |
