+++ framework/core/architecture/review-routes.json
@@ -1,6 +1,6 @@
 {
   "schema": "kungfu.core-review-routes/v1",
-  "authorityRoot": "sha256:7a88342cd221af5e808ea30c09e3705ee1807d5ff2c1a84e97be4aa17521ba3a",
+  "authorityRoot": "sha256:c36ce47601768cdd30c3ded5dd4e044fe425dc4482fccfe46208a26b7919d0a6",
   "fallbackAccount": "@dongkeren",
   "routes": [
     {
